+++ tests/common/cisco_data.py
@@ -86,3 +86,30 @@ def run_dshell_command(duthost, command):
     if not wait_until(300, 20, 0, check_dshell_ready, duthost):
         raise RuntimeError("Debug shell is not ready on {}".format(duthost.hostname))
     return duthost.shell(command)
+
+
+def copy_dshell_script_cisco_8000(dut, asic, dshell_script, script_name):
+    if dut.facts['asic_type'] != "cisco-8000":
+        raise RuntimeError("This function should have been called only for cisco-8000.")
+
+    script_path = "/tmp/{}".format(script_name)
+    dut.copy(content=dshell_script, dest=script_path)
+    if dut.sonichost.is_multi_asic:
+        dest = f"syncd{asic}"
+    else:
+        dest = "syncd"
+    dut.docker_copy_to_all_asics(
+        container_name=dest,
+        src=script_path,
+        dst="/")
+
+
+def copy_set_voq_watchdog_script_cisco_8000(dut, asic="", enable=True):
+    dshell_script = '''
+from common import d0
+def set_voq_watchdog(enable):
+    d0.set_bool_property(sdk.la_device_property_e_VOQ_WATCHDOG_ENABLED, enable)
+set_voq_watchdog({})
+'''.format(enable)
+
+    copy_dshell_script_cisco_8000(dut, asic, dshell_script, script_name="set_voq_watchdog.py")
